+++ l10n_br_hr/__init__.py
@@ -4,4 +4,3 @@
 # License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
 
 from . import model
-from . import wizard
