--- l10n_br_hr/wizard/__init__.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-# (c) 2019 Kmee - Diego Paradeda <diego.paradeda@kmee.com.br>
-# (c) 2019 Kmee - Hugo Borges <hugo.borges@kmee.com.br>
-# License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
-
-from . import hr_employee_timesheet_wizard
